backend/requirements.txt
flask
openai
python-dotenv
requests
pandas
